Update scripts in package.json and minor changes in gulpfile.ts and yarn.lock
The build script in package.json was modified to utilize a build shell script instead of the previous commands. This simplifies the build process and centralizes the build logic in a dedicated script. The gulpfile.ts file and yarn.lock have also been slightly modified, to match the updated build setup and dependencies.