Update dependencies in yarn.lock file
This commit updates the versions and dependencies within the yarn.lock file to keep the project's dependencies up-to-date. Included in the update are '@babel/code-frame@npm', 'async-each@npm', 'bn.js@npm' and several others. Updating these dependencies ensures the project stays current with the latest versions for improved functionality and security.