Move write-sys.sh script and update yarn.lock dependecies
The write-sys.sh script has been moved to a new 'scripts' directory. This better organizes the project structure and provides an appropriate location for related scripts. The execution flow of the write-sys.sh script has also been optimized. Moreover, dependencies in the yarn.lock file have been updated, which ensures we are using the most current and secure versions of our dependencies.